+++ src/lib/redirects.ts
@@ -1,0 +1,97 @@
+export const redirects = [
+	{
+		source: '/chat',
+		destination: 'https://cal.com/team/setangle/chat',
+		status: 308
+	},
+	{
+		source: '/agency',
+		destination: '/',
+		status: 308
+	},
+	{
+		source: '/zoom',
+		destination: 'https://cal.com/imsety/zoom/',
+		status: 308
+	},
+	{
+		source: '/motiontalent',
+		destination: 'https://tally.so/r/nW5b0J',
+		status: 308
+	},
+	{
+		source: '/goodtalent',
+		destination: 'https://tally.so/r/w8KKKO',
+		status: 308
+	},
+	{
+		source: '/consultation',
+		destination: 'https://cal.com/team/setangle/discuss',
+		status: 308
+	},
+	{
+		source: '/discuss',
+		destination: 'https://cal.com/team/setangle/discuss',
+		status: 308
+	},
+	{
+		source: '/talent',
+		destination: '/work-with-us',
+		status: 308
+	},
+	{
+		source: '/assistant',
+		destination: '/work-with-us/assistant',
+		status: 308
+	},
+	{
+		source: '/sanshu',
+		destination: '/work/sanshu',
+		status: 308
+	},
+	{
+		source: '/opencal',
+		destination: 'https://cal.com/team/setangle/opencal',
+		status: 308
+	},
+	{
+		source: '/goodpeople',
+		destination: '/work-with-us',
+		status: 308
+	},
+	{
+		source: '/interview',
+		destination: 'https://cal.com/team/setangle/interview',
+		status: 308
+	},
+	{
+		source: '/call',
+		destination: 'https://cal.com/team/setangle/call',
+		status: 308
+	},
+	{
+		source: '/talk',
+		destination: 'https://cal.com/team/setangle/talk',
+		status: 308
+	},
+	{
+		source: '/brief',
+		destination: 'https://cal.com/team/setangle/brief',
+		status: 308
+	},
+	{
+		source: '/abstract',
+		destination: 'https://www.instagram.com/p/C8XBcxbRgON/',
+		status: 308
+	},
+	{
+		source: '/sanshubehance',
+		destination: 'https://www.behance.net/gallery/202877441/SANSHU?share=1',
+		status: 308
+	},
+	{
+		source: '/connect',
+		destination: 'https://cal.com/team/setangle/connect',
+		status: 308
+	}
+];